OAF: Split off html and json rendering to own files
